Move customEditorInput to own file
