+++ src/vs/workbench/contrib/customEditor/browser/customEditors.ts
@@ -3,97 +3,25 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { memoize } from 'vs/base/common/decorators';
 import { UnownedDisposable } from 'vs/base/common/lifecycle';
-import { basename } from 'vs/base/common/path';
 import { endsWith } from 'vs/base/common/strings';
 import { withNullAsUndefined } from 'vs/base/common/types';
 import { URI } from 'vs/base/common/uri';
 import { generateUuid } from 'vs/base/common/uuid';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { IEditorModel, ITextEditorOptions } from 'vs/platform/editor/common/editor';
+import { ITextEditorOptions } from 'vs/platform/editor/common/editor';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { ILabelService } from 'vs/platform/label/common/label';
 import { IQuickInputService, IQuickPickItem } from 'vs/platform/quickinput/common/quickInput';
 import { IWorkbenchContribution } from 'vs/workbench/common/contributions';
-import { IEditor, IEditorInput, Verbosity } from 'vs/workbench/common/editor';
+import { IEditor, IEditorInput } from 'vs/workbench/common/editor';
 import { webviewEditorsExtensionPoint } from 'vs/workbench/contrib/customEditor/browser/extensionPoint';
 import { CustomEditorInfo, ICustomEditorService } from 'vs/workbench/contrib/customEditor/common/customEditor';
 import { FileEditorInput } from 'vs/workbench/contrib/files/common/editors/fileEditorInput';
 import { TEXT_FILE_EDITOR_ID } from 'vs/workbench/contrib/files/common/files';
-import { IWebviewService, WebviewEditorOverlay } from 'vs/workbench/contrib/webview/browser/webview';
-import { WebviewEditorInput } from 'vs/workbench/contrib/webview/browser/webviewEditorInput';
-import { IWebviewEditorService } from 'vs/workbench/contrib/webview/browser/webviewEditorService';
+import { IWebviewService } from 'vs/workbench/contrib/webview/browser/webview';
 import { IEditorGroup } from 'vs/workbench/services/editor/common/editorGroupsService';
 import { IEditorService, IOpenEditorOverride } from 'vs/workbench/services/editor/common/editorService';
-import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
-
-export class CustomFileEditorInput extends WebviewEditorInput {
-	private name?: string;
-
-	private _hasResolved = false;
-
-	constructor(
-		resource: URI,
-		viewType: string,
-		id: string,
-		webview: UnownedDisposable<WebviewEditorOverlay>,
-		@ILabelService private readonly labelService: ILabelService,
-		@IWebviewEditorService private readonly _webviewEditorService: IWebviewEditorService,
-		@IExtensionService private readonly _extensionService: IExtensionService,
-	) {
-		super(id, viewType, '', undefined, webview, resource);
-	}
-
-	getName(): string {
-		if (!this.name) {
-			this.name = basename(this.labelService.getUriLabel(this.editorResource));
-		}
-		return this.name;
-	}
-
-	matches(other: IEditorInput): boolean {
-		return this === other || (
-			other instanceof CustomFileEditorInput
-			&& this.viewType === other.viewType
-			&& this.editorResource.toString() === other.editorResource.toString());
-	}
-
-	@memoize
-	private get shortTitle(): string {
-		return this.getName();
-	}
-
-	@memoize
-	private get mediumTitle(): string {
-		return this.labelService.getUriLabel(this.editorResource, { relative: true });
-	}
-
-	@memoize
-	private get longTitle(): string {
-		return this.labelService.getUriLabel(this.editorResource);
-	}
-
-	getTitle(verbosity: Verbosity): string {
-		switch (verbosity) {
-			case Verbosity.SHORT:
-				return this.shortTitle;
-			default:
-			case Verbosity.MEDIUM:
-				return this.mediumTitle;
-			case Verbosity.LONG:
-				return this.longTitle;
-		}
-	}
-	public async resolve(): Promise<IEditorModel> {
-		if (!this._hasResolved) {
-			this._hasResolved = true;
-			this._extensionService.activateByEvent(`onWebviewEditor:${this.viewType}`);
-			await this._webviewEditorService.resolveWebview(this);
-		}
-		return super.resolve();
-	}
-}
+import { CustomFileEditorInput } from './customEditorInput';
 
 export class CustomEditorService implements ICustomEditorService {
 	_serviceBrand: any;
